Rewrote Keyboard interface from scratch.
Quite a big commit here. This commit begins a full rewrite of the
internal keyboard API, aiming to separate different sources of
keyboard input clearly and improve code clarity and usage experience.
A new dawn of the Keyboard abstraction layer.

The new Keyboard class is now a core that collects sources of input
(KeyboardSource's) and manages them in a centralized way. New sources
can be added inside the init code in main.ts. At the moment, there is
only one input source, which is the user's physical keyboard.

Helped-by: Strawberry <112416885+properly-strawberry@users.noreply.github.com>