Added functionality for the screen keyboard.
This is a pretty big commit again, it fixes many bugs introduced
while implementing this concept. In this commit, the on-screen
keyboard has been added as a keyboard source, represented by class
ScreenKeyboard. It tracks and reports the state of it. The class
also manages the modifiers on the keyboard, and tries to activate
them as global modifiers whenever possible; that is, whenever another
source is not making them active.

This commit also includes functionality to choose the right (modifier)
key, when a user remaps a modifier (e.g. CapsLock -> Control) on their
keyboard.