Fix CapsLock handling.
In this commit, it's been made so CapsLock, when pressed on the
physical keyboard, stays toggled until the next press, and
modifies the characters inputed after. The on-screen keyboard has
also been adjusted so CapsLock stays visually pressed while the
modifier is active.

This commit is really hacky in my opinion, and it is that because
browser event handling is cursed and terrible beyond repair.